requirements.txt
dash
pandas
numpy
google-cloud-bigquery==2.26.0
pandas-gbq==0.15.0
requests==2.26.0
google-api-python-client==2.20.0
gunicorn
plotly==5.3.1
pyarrow
